+++ src/events/lua_events/destroy_script_event.h
@@ -24,12 +24,12 @@
 namespace roa {
     struct destroy_script_event : public event_type {
         destroy_script_event() = default;
-        destroy_script_event(uint64_t id, uint64_t attached_entity_id) : event_type(destroy_script_event::type), id(id), attached_entity_id(attached_entity_id) {}
+        destroy_script_event(uint32_t id, uint64_t attached_entity_id) : event_type(destroy_script_event::type), id(id), attached_entity_id(attached_entity_id) {}
         virtual ~destroy_script_event() {}
 
         static constexpr uint32_t type = 2;
-        uint64_t id;
-        uint64_t attached_entity_id;
+        uint32_t id;
+        uint32_t attached_entity_id;
     };
 }
